#1476 refactor ComposeViewController
