+++ FlowCrypt/Controllers/Compose/Extensions/ComposeViewController+Keyboard.swift
@@ -1,0 +1,58 @@
+//
+//  ComposeViewController+Keyboard.swift
+//  FlowCrypt
+//
+//  Created by Ioan Moldovan on 4/6/22
+//  Copyright © 2017-present FlowCrypt a. s. All rights reserved.
+//
+
+import UIKit
+import FlowCryptUI
+
+// MARK: - Keyboard
+extension ComposeViewController {
+    internal func observeKeyboardNotifications() {
+        // swiftlint:disable discarded_notification_center_observer
+        NotificationCenter.default.addObserver(
+            forName: UIResponder.keyboardWillShowNotification,
+            object: nil,
+            queue: .main
+        ) { [weak self] notification in
+            guard let self = self else { return }
+            self.adjustForKeyboard(height: self.keyboardHeight(from: notification))
+        }
+
+        NotificationCenter.default.addObserver(
+            forName: UIResponder.keyboardWillHideNotification,
+            object: nil,
+            queue: .main
+        ) { [weak self] _ in
+            self?.adjustForKeyboard(height: 0)
+        }
+    }
+
+    internal func observerAppStates() {
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(startDraftTimer),
+            name: UIApplication.didBecomeActiveNotification,
+            object: nil)
+
+        NotificationCenter.default.addObserver(
+            self,
+            selector: #selector(stopDraftTimer),
+            name: UIApplication.willResignActiveNotification,
+            object: nil)
+    }
+
+    internal func adjustForKeyboard(height: CGFloat) {
+        node.contentInset.bottom = height + 8
+
+        guard let textView = node.visibleNodes.compactMap({ $0 as? TextViewCellNode }).first?.textView.textView,
+              let selectedRange = textView.selectedTextRange
+        else { return }
+
+        let rect = textView.caretRect(for: selectedRange.start)
+        node.view.scrollRectToVisible(rect, animated: true)
+    }
+}
