+++ FlowCrypt/Controllers/Compose/Extensions/ComposeViewController+Search.swift
@@ -1,0 +1,27 @@
+//
+//  ComposeViewController+Search.swift
+//  FlowCrypt
+//
+//  Created by Ioan Moldovan on 4/6/22
+//  Copyright © 2017-present FlowCrypt a. s. All rights reserved.
+//
+
+// MARK: - Search
+extension ComposeViewController {
+    internal func setupSearch() {
+        search
+            .debounce(for: .milliseconds(300), scheduler: RunLoop.main)
+            .removeDuplicates()
+            .map { [weak self] query -> String in
+                if query.isEmpty {
+                    self?.updateState(with: .main)
+                }
+                return query
+            }
+            .sink(receiveValue: { [weak self] in
+                guard $0.isNotEmpty else { return }
+                self?.searchEmail(with: $0)
+            })
+            .store(in: &cancellable)
+    }
+}
